Use linux parsing rather than Solaris for IP stats
