+++ oshi-core/src/main/java/oshi/software/os/linux/LinuxInternetProtocolStats.java
@@ -43,8 +43,8 @@ import java.util.Map;
 
 import oshi.annotation.concurrent.ThreadSafe;
 import oshi.driver.linux.proc.ProcessStat;
+import oshi.driver.unix.NetStat;
 import oshi.software.common.AbstractInternetProtocolStats;
-import oshi.util.ExecutingCommand;
 import oshi.util.FileUtil;
 import oshi.util.ParseUtil;
 import oshi.util.platform.linux.ProcPath;
@@ -55,7 +55,7 @@ public class LinuxInternetProtocolStats extends AbstractInternetProtocolStats {
 
     @Override
     public TcpStats getTCPv4Stats() {
-        return getTcpStats("netstat -st4");
+        return NetStat.queryTcpStats("netstat -st4");
     }
 
     @Override
@@ -64,100 +64,14 @@ public class LinuxInternetProtocolStats extends AbstractInternetProtocolStats {
         return new TcpStats(0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L, 0L);
     }
 
-    private static TcpStats getTcpStats(String netstatStr) {
-        long connectionsEstablished = 0;
-        long connectionsActive = 0;
-        long connectionsPassive = 0;
-        long connectionFailures = 0;
-        long connectionsReset = 0;
-        long segmentsSent = 0;
-        long segmentsReceived = 0;
-        long segmentsRetransmitted = 0;
-        long inErrors = 0;
-        long outResets = 0;
-        List<String> netstat = ExecutingCommand.runNative(netstatStr);
-        for (String s : netstat) {
-            String[] split = s.trim().split(" ", 2);
-            if (split.length == 2) {
-                switch (split[1]) {
-                case "connections established":
-                    connectionsEstablished = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "active connection openings":
-                    connectionsActive = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "passive connection openings":
-                    connectionsPassive = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "failed connection attempts":
-                    connectionFailures = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "connection resets received":
-                    connectionsReset = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "segments sent out":
-                    segmentsSent = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "segments received":
-                    segmentsReceived = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "segments retransmitted":
-                    segmentsRetransmitted = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "bad segments received":
-                    inErrors = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "resets sent":
-                    outResets = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                default:
-                    break;
-                }
-
-            }
-        }
-        return new TcpStats(connectionsEstablished, connectionsActive, connectionsPassive, connectionFailures,
-                connectionsReset, segmentsSent, segmentsReceived, segmentsRetransmitted, inErrors, outResets);
-    }
-
     @Override
     public UdpStats getUDPv4Stats() {
-        return getUdpStats("netstat -su4");
+        return NetStat.queryUdpStats("netstat -su4");
     }
 
     @Override
     public UdpStats getUDPv6Stats() {
-        return getUdpStats("netstat -su6");
-    }
-
-    private static UdpStats getUdpStats(String netstatStr) {
-        long datagramsSent = 0;
-        long datagramsReceived = 0;
-        long datagramsNoPort = 0;
-        long datagramsReceivedErrors = 0;
-        List<String> netstat = ExecutingCommand.runNative(netstatStr);
-        for (String s : netstat) {
-            String[] split = s.trim().split(" ", 2);
-            if (split.length == 2) {
-                switch (split[1]) {
-                case "packets sent":
-                    datagramsSent = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "packets received":
-                    datagramsReceived = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "packets to unknown port received":
-                    datagramsNoPort = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                case "packet receive errors":
-                    datagramsReceivedErrors = ParseUtil.parseLongOrDefault(split[0], 0L);
-                    break;
-                default:
-                    break;
-                }
-            }
-        }
-        return new UdpStats(datagramsSent, datagramsReceived, datagramsNoPort, datagramsReceivedErrors);
+        return NetStat.queryUdpStats("netstat -su6");
     }
 
     @Override
